Move decorators and yaml validation logic into clowder_repo
