+++ clowder/clowder/cli/forall_controller.py
@@ -8,11 +8,11 @@
 from cement.ext.ext_argparse import ArgparseController, expose
 
 import clowder.commands as commands
-from clowder.cli.globals import CLOWDER_CONTROLLER
-from clowder.util.decorators import (
+from clowder.clowder_repo import (
     print_clowder_repo_status,
     valid_clowder_yaml_required
 )
+from clowder.cli.globals import CLOWDER_CONTROLLER
 
 
 class ForallController(ArgparseController):
