+++ clowder/clowder/cli/globals.py
@@ -8,8 +8,6 @@
 import os
 
 from clowder.clowder_controller import ClowderController
-from clowder.clowder_repo import ClowderRepo
 
 
 CLOWDER_CONTROLLER = ClowderController(os.getcwd())
-CLOWDER_REPO = ClowderRepo(os.getcwd())
