+++ clowder/clowder/cli/prune_controller.py
@@ -10,6 +10,10 @@ import sys
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
+from clowder.clowder_repo import (
+    print_clowder_repo_status,
+    valid_clowder_yaml_required
+)
 from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.commands.util import (
     existing_branch_groups,
@@ -21,11 +25,7 @@ from clowder.commands.util import (
     validate_groups,
     validate_projects
 )
-from clowder.util.decorators import (
-    print_clowder_repo_status,
-    network_connection_required,
-    valid_clowder_yaml_required
-)
+from clowder.util.decorators import network_connection_required
 
 
 class PruneController(ArgparseController):
