+++ clowder/clowder/cli/save_controller.py
@@ -11,11 +11,11 @@ import sys
 from cement.ext.ext_argparse import ArgparseController, expose
 
 import clowder.util.formatting as fmt
-from clowder.cli.globals import (
-    CLOWDER_CONTROLLER,
-    CLOWDER_REPO
+from clowder.clowder_repo import (
+    CLOWDER_REPO,
+    valid_clowder_yaml_required
 )
-from clowder.util.decorators import valid_clowder_yaml_required
+from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.commands.util import (
     validate_groups,
     validate_projects_exist
