+++ clowder/clowder/cli/stash_controller.py
@@ -7,16 +7,16 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
+from clowder.clowder_repo import (
+    print_clowder_repo_status,
+    valid_clowder_yaml_required
+)
 from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.commands.util import (
     filter_groups,
     filter_projects_on_project_names,
     run_group_command,
     run_project_command
-)
-from clowder.util.decorators import (
-    print_clowder_repo_status,
-    valid_clowder_yaml_required
 )
