+++ clowder/clowder/cli/status_controller.py
@@ -5,13 +5,13 @@
 
 """
 
+import os
+
 from cement.ext.ext_argparse import ArgparseController, expose
 
 import clowder.util.formatting as fmt
-from clowder.cli.globals import (
-    CLOWDER_CONTROLLER,
-    CLOWDER_REPO
-)
+from clowder.clowder_repo import CLOWDER_REPO
+from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.commands.util import run_group_command
 from clowder.util.decorators import network_connection_required
