+++ clowder/clowder/cli/sync_controller.py
@@ -12,13 +12,13 @@ from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
 import clowder.commands as commands
-from clowder.cli.globals import CLOWDER_CONTROLLER
-from clowder.cli.util import options_help_message
-from clowder.util.decorators import (
+from clowder.clowder_repo import (
     print_clowder_repo_status_fetch,
-    network_connection_required,
     valid_clowder_yaml_required
 )
+from clowder.cli.globals import CLOWDER_CONTROLLER
+from clowder.cli.util import options_help_message
+from clowder.util.decorators import network_connection_required
 
 
 class SyncController(ArgparseController):
