+++ clowder/clowder/cli/yaml_controller.py
@@ -11,12 +11,12 @@ import sys
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.cli.globals import CLOWDER_CONTROLLER
 import clowder.util.formatting as fmt
-from clowder.util.decorators import (
+from clowder.clowder_repo import (
     print_clowder_repo_status,
     valid_clowder_yaml_required
 )
+from clowder.cli.globals import CLOWDER_CONTROLLER
 from clowder.yaml.printing import print_yaml
