+++ clowder/clowder/clowder_repo.py
@@ -12,13 +12,15 @@ import errno
 import os
 import sys
 
-from termcolor import colored
+from termcolor import colored, cprint
 
 import clowder.util.formatting as fmt
+from clowder.error.clowder_error import ClowderError
 from clowder.git.project_repo import ProjectRepo
 from clowder.util.connectivity import is_offline
 from clowder.util.execute import execute_command
 from clowder.util.file_system import remove_directory
+from clowder.yaml.validating import validate_yaml
 
 
 class ClowderRepo(object):
@@ -42,10 +44,22 @@ class ClowderRepo(object):
         self.clowder_path = os.path.join(self.root_directory, '.clowder')
 
         # Create clowder.yaml symlink if .clowder dir and yaml file exist
+        clowder_symlink = os.path.join(self.root_directory, 'clowder.yaml')
         if os.path.isdir(self.clowder_path):
-            clowder_symlink = os.path.join(self.root_directory, 'clowder.yaml')
             if not os.path.islink(clowder_symlink):
                 self.link()
+
+        self.is_yaml_valid = False
+        self.error = None
+        if os.path.islink(clowder_symlink):
+            try:
+                validate_yaml(os.path.join(self.root_directory, 'clowder.yaml'), self.root_directory)
+            except ClowderError as err:
+                self.error = err
+            except (KeyboardInterrupt, SystemExit):
+                sys.exit(1)
+            else:
+                self.is_yaml_valid = True
 
     def add(self, files):
         """Add files in clowder repo to git index
@@ -218,6 +232,21 @@ class ClowderRepo(object):
             sys.exit(1)
 
 
+CLOWDER_REPO = ClowderRepo(os.getcwd())
+
+
+def clowder_required(func):
+    """If no clowder repo, print clowder not found message and exit"""
+
+    def wrapper(*args, **kwargs):
+        """Wrapper"""
+
+        _validate_clowder_repo_exists(CLOWDER_REPO)
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
 def force_symlink(file1, file2):
     """Force symlink creation
 
@@ -234,4 +263,55 @@ def force_symlink(file1, file2):
     except (KeyboardInterrupt, SystemExit):
         os.remove(file2)
         os.symlink(file1, file2)
+        sys.exit(1)
+
+
+def print_clowder_repo_status(func):
+    """Print clowder repo status"""
+
+    def wrapper(*args, **kwargs):
+        """Wrapper"""
+
+        CLOWDER_REPO.print_status()
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
+def print_clowder_repo_status_fetch(func):
+    """Print clowder repo status"""
+
+    def wrapper(*args, **kwargs):
+        """Wrapper"""
+
+        CLOWDER_REPO.print_status(fetch=True)
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
+def valid_clowder_yaml_required(func):
+    """If clowder.yaml is invalid, print invalid yaml message and exit"""
+
+    def wrapper(*args, **kwargs):
+        """Wrapper"""
+
+        _validate_clowder_repo_exists(CLOWDER_REPO)
+        if not CLOWDER_REPO.is_yaml_valid:
+            print(fmt.invalid_yaml_error())
+            print(fmt.error(CLOWDER_REPO.error))
+            sys.exit(1)
+        return func(*args, **kwargs)
+
+    return wrapper
+
+
+def _validate_clowder_repo_exists(repo):
+    """If clowder repo doesn't exist, print message and exit
+
+    :param ClowderRepo repo: Repo to check
+    """
+
+    if not os.path.isdir(CLOWDER_REPO.clowder_path):
+        cprint(' - No .clowder found in the current directory\n', 'red')
         sys.exit(1)
