+++ clowder/clowder/util/decorators.py
@@ -11,20 +11,7 @@ import sys
 from termcolor import cprint
 
 import clowder.util.formatting as fmt
-from clowder.cli.globals import CLOWDER_REPO
 from clowder.util.connectivity import is_offline
-
-
-def clowder_required(func):
-    """If no clowder repo, print clowder not found message and exit"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        _validate_clowder_repo_exists(CLOWDER_REPO)
-        return func(*args, **kwargs)
-
-    return wrapper
 
 
 def not_detached(func):
@@ -56,30 +43,6 @@ def project_repo_exists(func):
     return wrapper
 
 
-def print_clowder_repo_status(func):
-    """Print clowder repo status"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        CLOWDER_REPO.print_status()
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def print_clowder_repo_status_fetch(func):
-    """Print clowder repo status"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        CLOWDER_REPO.print_status(fetch=True)
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
 def network_connection_required(func):
     """If no network connection, print offline message and exit"""
 
@@ -92,31 +55,3 @@ def network_connection_required(func):
         return func(*args, **kwargs)
 
     return wrapper
-
-
-def valid_clowder_yaml_required(func):
-    """If clowder.yaml is invalid, print invalid yaml message and exit"""
-
-    def wrapper(*args, **kwargs):
-        """Wrapper"""
-
-        instance = args[0]
-        _validate_clowder_repo_exists(instance.clowder_repo)
-        if instance.invalid_yaml:
-            print(fmt.invalid_yaml_error())
-            print(fmt.error(instance.error))
-            sys.exit(1)
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def _validate_clowder_repo_exists(repo):
-    """If clowder repo doesn't exist, print message and exit
-
-    :param ClowderRepo repo: Repo to check
-    """
-
-    if repo is None:
-        cprint(' - No clowder found in the current directory\n', 'red')
-        sys.exit(1)
